Group .kt files in `protobuf` module under its dedicated package name
