+++ protobuf/src/main/kotlin/io/spine/chords/protobuf/form/InputFieldExts.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.chords.form
+package io.spine.chords.protobuf.form
 
 import com.google.protobuf.Message
 import io.spine.chords.InputField
