+++ protobuf/src/main/kotlin/io/spine/chords/protobuf/form/MessageFormScopes.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.chords.form
+package io.spine.chords.protobuf.form
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
