+++ protobuf/src/main/kotlin/io/spine/chords/protobuf/net/EmailField.kt
@@ -1,0 +1,75 @@
+/*
+ * Copyright 2024, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.chords.protobuf.net
+
+import io.spine.chords.ComponentCompanion
+import io.spine.chords.InputField
+import io.spine.chords.ValueParseException
+import io.spine.net.EmailAddress
+
+/**
+ * Regex pattern for the email address.
+ */
+private const val EmailRegex = """^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}${'$'}"""
+
+/**
+ * An input field that allows editing a [EmailAddress] field.
+ */
+public class EmailField : InputField<EmailAddress>() {
+    public companion object : ComponentCompanion<EmailField>({ EmailField() })
+
+    init {
+        label = "Email"
+        promptText = "example@gmail.com"
+    }
+
+    override fun parseValue(rawText: String): EmailAddress {
+        validate(rawText)
+        return EmailAddress
+            .newBuilder()
+            .setValue(rawText)
+            .vBuild()
+    }
+
+    override fun formatValue(value: EmailAddress): String {
+        return value.value
+    }
+
+    /**
+     * Validate the email address value.
+     *
+     * @throws ValueParseException
+     *         if the provided email value does not match the [EmailRegex] pattern.
+     */
+    @Throws(ValueParseException::class)
+    private fun validate(email: String) {
+        val emailPattern = Regex(EmailRegex)
+        if (!emailPattern.matches(email)) {
+            throw ValueParseException("Enter a valid email address.")
+        }
+    }
+}
