+++ protobuf/src/main/kotlin/io/spine/chords/protobuf/time/TimeExts.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.chords.time
+package io.spine.chords.protobuf.time
 
 import com.google.protobuf.Duration
 import com.google.protobuf.Timestamp
