+++ protobuf/src/test/kotlin/io/spine/chords/money/MoneyFieldSpec.kt
@@ -31,7 +31,7 @@ import io.kotest.assertions.withClue
 import io.kotest.core.spec.DisplayName
 import io.kotest.matchers.shouldBe
 import io.spine.chords.RawTextContent
-import io.spine.chords.money.MoneyFieldReviser
+import io.spine.chords.protobuf.money.MoneyFieldReviser
 import io.spine.money.Currency
 import io.spine.money.Currency.BIF
 import io.spine.money.Currency.IQD
